backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.1
pymysql==1.1.1
cryptography==42.0.8
boto3==1.34.79
python-dotenv==1.0.1
gunicorn==22.0.0
